separating cors to a configuration file
