+++ src/main/java/dev/gerardo/shippingapp/ShippingAppApplication.java
@@ -13,16 +13,4 @@ public class ShippingAppApplication {
 		SpringApplication.run(ShippingAppApplication.class, args);
 	}
 
-	@Bean
-	public WebMvcConfigurer corsConfigurer() {
-		return new WebMvcConfigurer() {
-			@Override
-			public void addCorsMappings(CorsRegistry registry) {
-				registry.addMapping("/type")
-						.allowedMethods("GET")
-						.allowedOrigins("*");
-			}
-		};
-	}
-
 }
